chatbot/requirements.txt
# Core dependencies for Hue Chatbot
google-generativeai>=0.3.0
duckduckgo-search>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0